[Refactor] Update some rule's interface
